AI/Notebooklm_lite/requirements.txt
streamlit==1.45.1
python-dotenv==1.2.2
langchain==1.2.18
langchain-core==1.3.3
langchain-text-splitters==1.1.2
langchain-openai==1.2.1
langchain-community==0.4.1
chromadb==1.5.9
pypdf==6.11.0
